Plot world generators - Add functionality for custom plot world generators - Implement plot bounds checking that relies on a generator's configured plot and padding sizes rather than a hard-coded 1x1 grid - Add ability to register custom plot world generators - Add 4 default plot world generators `plains1-4` - Plot worlds are now automatically loaded upon plugin enable - Add configuration option `defaultPlotWorldGenerator` - Add admin sub-command `listgen` to list all registered plot world generators - Update admin sub-command `genworld` to optionally accept a generator - Update admin sub-command `import` to accept a generator - Fix general sub-command `tp` continuing to execute after receiving a non-plot world - Add check to general sub-command `info` which checks to see if the plot has an invalid world ID - Remove debug plot world chunk generator - Replace Checker Framework @Nullable with Jetbrains @Nullable to be consistent with Bukkit API

+++ src/main/java/me/whizvox/infiniplots/db/PlotRepository.java
@@ -3,7 +3,7 @@ package me.whizvox.infiniplots.db;
 import me.whizvox.infiniplots.InfiniPlots;
 import me.whizvox.infiniplots.plot.Plot;
 import me.whizvox.infiniplots.util.ChunkPos;
-import org.checkerframework.checker.nullness.qual.Nullable;
+import org.jetbrains.annotations.Nullable;
 
 import java.sql.Connection;
 import java.sql.ResultSet;
@@ -105,9 +105,11 @@ public class PlotRepository extends Repository {
 
   public void updateOwner(UUID plotId, UUID newOwner) {
     Plot plot = getById(plotId, false);
-    executeUpdate(UPDATE_OWNER, List.of(newOwner, plotId));
-    memberRepo.removeMember(plotId, newOwner);
-    memberRepo.addMember(plotId, plot.owner());
+    if (plot != null) {
+      executeUpdate(UPDATE_OWNER, List.of(newOwner, plotId));
+      memberRepo.removeMember(plotId, newOwner);
+      memberRepo.addMember(plotId, plot.owner());
+    }
   }
 
   public void remove(UUID id) {
